general: save memory by replacing s32 for u8

+++ board.h
@@ -1,15 +1,16 @@
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdint.h>
 #include "boat.h"
 
 
 typedef struct
 {
-  int   height;
-  int   width;
-  int** matrix;
+  uint8_t   height;
+  uint8_t   width;
+  uint8_t** matrix;
 } Board;
 
-Board* construct_board(int height, int width);
+Board* construct_board(uint8_t height, uint8_t width);
 void destruct_board(Board* board);
 void print_board(Board* board);
