+++ game.h
@@ -1,6 +1,7 @@
 #include <stdbool.h>
 #include <stdlib.h>
 #include <stdio.h>
+#include <stdint.h>
 #include "board.h"
 
 
@@ -12,6 +13,6 @@ typedef struct
   Board* board_p2;
 } Game;
 
-Game* construct_game(int height, int width, int* boats);
+Game* construct_game(uint8_t height, uint8_t width, int* boats);
 void destruct_game(Game* game);
 void print_game(Game* game);
